requirements.txt
numpy==2.2.4
pillow==11.1.0
unicornpihd
